requirements.txt
flask==3.0.0
flask-cors==4.0.0
waitress==2.1.2
python-dotenv==1.0.0